fix(navigation): hide index tab while preserving redirect functionality
- Add index route to handle default tab navigation
- Configure tab layout to hide index from tab bar
- Update navigation paths to be more explicit
- Maintain navigation state on auth status change

Resolves navigation issues after login and removes unwanted index tab from UI
